refactor(store): split sessionsSlice into crud/runtime/titleBackfill (Task #736)
Per PR #754 review suggestion. Split the 530-LoC `sessionsSlice.ts`
into three SRP-focused sub-slices:

- `sessionCrudSlice` (404 LoC): create / import / delete / restore /
  move / rename / changeCwd / setSessionModel + active selection +
  LRU cwd seed. Hosts `userHome` / `claudeSettingsDefaultModel` /
  `lastUsedCwd` initial state and `ensureUsableGroup` helper.
- `sessionRuntimeSlice` (100 LoC): transient runtime state owned by
  the SDK stream / pty bridges. `_applySessionState`, `_setFlash`,
  `_applyCwdRedirect`, `_applyPtyExit`, `_clearPtyExit`. Owns
  `flashStates` + `disconnectedSessions` initial state.
- `sessionTitleBackfillSlice` (88 LoC): JSONL-derived title sync.
  `_applyExternalTitle` (raw name patch primitive) +
  `_backfillTitles` (one-shot post-hydrate pull from
  `ccsmSessionTitles.listForProject`).

`store.ts` now spreads the three new slices in place of the old
`createSessionsSlice`. The old file is deleted; one stale doc-comment
in `modelPickerSlice.ts` was updated to reference `sessionCrudSlice`.

Public store API is unchanged — `RootStore` shape (State + Actions)
in `slices/types.ts` was not touched, so every consumer (`useStore`,
hydrate path, IPC bridges, components) continues to work without
import edits. Verified by all 5 `tests/store-*.test.ts` suites
passing with zero changes.

Test split:
- `tests/stores/slices/sessionsSlice.test.ts` (308 LoC, 27 tests)
  → `sessionCrudSlice.test.ts` (17 tests),
    `sessionRuntimeSlice.test.ts` (5 tests),
    `sessionTitleBackfillSlice.test.ts` (3 tests).

Validation:
- `npx tsc --noEmit` clean.
- `npx vitest run tests/stores/` 67/67 pass (7 files).
- `npx vitest run tests/store-*.test.ts` 33/33 pass (proves public API
  unchanged).
- `npm run build` clean (3 pre-existing webpack size warnings).
- E2E `harness-real-cli --only=new-session-chat` PASS (21.6s).
- Reverse-verify on `sessionRuntimeSlice`: stash file → 1 test FAIL
  (import resolves to nothing); restore → 5/5 PASS.
